pyproject.toml
[tool.poetry]
name = "secret weapon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = "0.21.1"
bcrypt = "4.0.1"
sqlalchemy = "2.0.10"
python-jose = "3.3.0"
passlib = "1.7.4"
python-multipart = "0.0.6"
requests = "2.31.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
